Stop internal hamlib install running in main install phase
Because  the hamlib-update  phase always  runs it  was triggering  the
hamlib-configure,  hamlib-build,   and  hamlib-install   phases  every
time. This  caused an issue  when the  install phase is  running under
fakeroot on a build server as  the halib-install phase would install a
location  under  the  fake  root rather  than  the  intended  internal
directory in the build tree.

Simply skipping the  hamlib-update phase proves to be  the solution as
it breaks the  whole unwanted chain. This assumes  that hamlib updates
are never going to be picked up on rebuilds which we can probably live
with.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-superbuild@4508 ab8295b8-cf94-4d9e-aec4-7959e3be5d79