Optimize the WSJT-X superbuild source target
The source target  need not build the WSJT-X external  project as only
up to the configure step is required.

The hamlib external project must still complete up to the install step
as it is needed to configure the WSJT-X external project.

This  reduces  the build  time  of  this  target and  doesn't  require
asciidoc since  it doesn't have to  generate the manpages to  make the
source tarball. To achieve this it  is required to pass a CMake option
at configure time:

   cmake -D WSJTX_EXTRA_CMAKE_OPTIONS="-D;WSJT_SKIP_MANPAGES=ON" <source-path>

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-superbuild@4570 ab8295b8-cf94-4d9e-aec4-7959e3be5d79